Lichtsensor main projectoplevering ready
